Extract execution state from runner-state into a new runner-execution-core crate
