+++ Cargo.toml
@@ -37,6 +37,7 @@ waymark-prometheus-exporter-bringup = { path = "crates/lib/prometheus-exporter-b
 waymark-proto = { path = "crates/lib/proto" }
 waymark-runloop = { path = "crates/lib/runloop" }
 waymark-runner = { path = "crates/lib/runner" }
+waymark-runner-execution-core = { path = "crates/lib/runner-execution-core" }
 waymark-runner-executor-core = { path = "crates/lib/runner-executor-core" }
 waymark-runner-expr = { path = "crates/lib/runner-expr" }
 waymark-runner-expr-eval = { path = "crates/lib/runner-expr-eval" }
