+++ crates/lib/core-backend/Cargo.toml
@@ -10,6 +10,7 @@ default = ["either"]
 [dependencies]
 waymark-backends-core = { workspace = true }
 waymark-ids = { workspace = true }
+waymark-runner-execution-core = { workspace = true }
 waymark-runner-executor-core = { workspace = true }
 waymark-runner-state = { workspace = true }
