+++ crates/lib/runner-state/src/lib.rs
@@ -1,6 +1,5 @@
 mod collect_value_sources;
 mod error;
-mod execution;
 mod max_nodes;
 mod resolve_value_tree;
 mod state;
@@ -8,7 +7,6 @@ mod util;
 mod value;
 
 pub use self::error::*;
-pub use self::execution::*;
 pub use self::state::*;
 pub use self::value::*;
