+++ crates/lib/runner/Cargo.toml
@@ -11,6 +11,7 @@ waymark-ids = { workspace = true }
 waymark-ir-conversions = { workspace = true }
 waymark-observability = { workspace = true }
 waymark-proto = { workspace = true }
+waymark-runner-execution-core = { workspace = true }
 waymark-runner-executor-core = { workspace = true }
 waymark-runner-expr-eval = { workspace = true }
 waymark-runner-retry-policy = { workspace = true }
